+++ static/testdata/handlers/dynamic-routes/projection.go
@@ -26,35 +26,32 @@ func (ProjectionHandler) Configure(c dogma.ProjectionConfigurer) {
 
 // HandleEvent updates the projection to reflect the occurrence of an event.
 func (ProjectionHandler) HandleEvent(
-	ctx context.Context,
-	r, c, n []byte,
-	s dogma.ProjectionEventScope,
-	m dogma.Message,
+	_ context.Context,
+	_, _, _ []byte,
+	_ dogma.ProjectionEventScope,
+	_ dogma.Event,
 ) (ok bool, err error) {
 	return false, nil
 }
 
 // ResourceVersion returns the version of the resource r.
-func (ProjectionHandler) ResourceVersion(
-	ctx context.Context,
-	r []byte,
-) ([]byte, error) {
+func (ProjectionHandler) ResourceVersion(context.Context, []byte) ([]byte, error) {
 	return nil, nil
 }
 
 // CloseResource informs the projection that the resource r will not be
 // used in any future calls to HandleEvent().
-func (ProjectionHandler) CloseResource(ctx context.Context, r []byte) error {
+func (ProjectionHandler) CloseResource(context.Context, []byte) error {
 	return nil
 }
 
 // TimeoutHint returns a duration that is suitable for computing a deadline
 // for the handling of the given message by this handler.
-func (ProjectionHandler) TimeoutHint(m dogma.Message) time.Duration {
+func (ProjectionHandler) TimeoutHint(dogma.Message) time.Duration {
 	return 0
 }
 
 // Compact reduces the size of the projection's data.
-func (ProjectionHandler) Compact(ctx context.Context, s dogma.ProjectionCompactScope) error {
+func (ProjectionHandler) Compact(context.Context, dogma.ProjectionCompactScope) error {
 	return nil
 }