+++ static/testdata/handlers/multiple/first.go
@@ -13,7 +13,7 @@ type FirstAggregate struct{}
 
 // ApplyEvent updates the aggregate instance to reflect the occurrence of an
 // event that was recorded against this instance.
-func (FirstAggregate) ApplyEvent(m dogma.Message) {}
+func (FirstAggregate) ApplyEvent(dogma.Event) {}
 
 // FirstAggregateHandler is a test implementation of dogma.AggregateMessageHandler.
 type FirstAggregateHandler struct{}
@@ -36,15 +36,15 @@ func (FirstAggregateHandler) Configure(c dogma.AggregateConfigurer) {
 
 // RouteCommandToInstance returns the ID of the aggregate instance that is
 // targetted by m.
-func (FirstAggregateHandler) RouteCommandToInstance(m dogma.Message) string {
+func (FirstAggregateHandler) RouteCommandToInstance(dogma.Command) string {
 	return "<first-aggregate>"
 }
 
 // HandleCommand handles a command message that has been routed to this handler.
 func (FirstAggregateHandler) HandleCommand(
-	r dogma.AggregateRoot,
-	s dogma.AggregateCommandScope,
-	m dogma.Message,
+	dogma.AggregateRoot,
+	dogma.AggregateCommandScope,
+	dogma.Command,
 ) {
 }
 
@@ -75,18 +75,18 @@ func (FirstProcessHandler) Configure(c dogma.ProcessConfigurer) {
 // RouteEventToInstance returns the ID of the process instance that is
 // targeted by m.
 func (FirstProcessHandler) RouteEventToInstance(
-	ctx context.Context,
-	m dogma.Message,
+	context.Context,
+	dogma.Event,
 ) (string, bool, error) {
 	return "<first-process>", true, nil
 }
 
 // HandleEvent handles an event message.
 func (FirstProcessHandler) HandleEvent(
-	ctx context.Context,
-	r dogma.ProcessRoot,
-	s dogma.ProcessEventScope,
-	m dogma.Message,
+	context.Context,
+	dogma.ProcessRoot,
+	dogma.ProcessEventScope,
+	dogma.Event,
 ) error {
 	return nil
 }
@@ -94,17 +94,17 @@ func (FirstProcessHandler) HandleEvent(
 // HandleTimeout handles a timeout message that has been scheduled with
 // ProcessScope.ScheduleTimeout().
 func (FirstProcessHandler) HandleTimeout(
-	ctx context.Context,
-	r dogma.ProcessRoot,
-	s dogma.ProcessTimeoutScope,
-	m dogma.Message,
+	context.Context,
+	dogma.ProcessRoot,
+	dogma.ProcessTimeoutScope,
+	dogma.Timeout,
 ) error {
 	return nil
 }
 
 // TimeoutHint returns a duration that is suitable for computing a deadline
 // for the handling of the given message by this handler.
-func (FirstProcessHandler) TimeoutHint(m dogma.Message) time.Duration {
+func (FirstProcessHandler) TimeoutHint(dogma.Message) time.Duration {
 	return 0
 }
 
@@ -124,36 +124,33 @@ func (FirstProjectionHandler) Configure(c dogma.ProjectionConfigurer) {
 
 // HandleEvent updates the projection to reflect the occurrence of an event.
 func (FirstProjectionHandler) HandleEvent(
-	ctx context.Context,
-	r, c, n []byte,
-	s dogma.ProjectionEventScope,
-	m dogma.Message,
+	_ context.Context,
+	_, _, _ []byte,
+	_ dogma.ProjectionEventScope,
+	_ dogma.Event,
 ) (ok bool, err error) {
 	return false, nil
 }
 
 // ResourceVersion returns the version of the resource r.
-func (FirstProjectionHandler) ResourceVersion(
-	ctx context.Context,
-	r []byte,
-) ([]byte, error) {
+func (FirstProjectionHandler) ResourceVersion(context.Context, []byte) ([]byte, error) {
 	return nil, nil
 }
 
 // CloseResource informs the projection that the resource r will not be
 // used in any future calls to HandleEvent().
-func (FirstProjectionHandler) CloseResource(ctx context.Context, r []byte) error {
+func (FirstProjectionHandler) CloseResource(context.Context, []byte) error {
 	return nil
 }
 
 // TimeoutHint returns a duration that is suitable for computing a deadline
 // for the handling of the given message by this handler.
-func (FirstProjectionHandler) TimeoutHint(m dogma.Message) time.Duration {
+func (FirstProjectionHandler) TimeoutHint(dogma.Message) time.Duration {
 	return 0
 }
 
 // Compact reduces the size of the projection's data.
-func (FirstProjectionHandler) Compact(ctx context.Context, s dogma.ProjectionCompactScope) error {
+func (FirstProjectionHandler) Compact(context.Context, dogma.ProjectionCompactScope) error {
 	return nil
 }
 
@@ -174,21 +171,21 @@ func (FirstIntegrationHandler) Configure(c dogma.IntegrationConfigurer) {
 
 // RouteCommandToInstance returns the ID of the integration instance that is
 // targetted by m.
-func (FirstIntegrationHandler) RouteCommandToInstance(m dogma.Message) string {
+func (FirstIntegrationHandler) RouteCommandToInstance(dogma.Command) string {
 	return "<first-integration>"
 }
 
 // HandleCommand handles a command message that has been routed to this handler.
 func (FirstIntegrationHandler) HandleCommand(
-	ctx context.Context,
-	s dogma.IntegrationCommandScope,
-	m dogma.Message,
+	context.Context,
+	dogma.IntegrationCommandScope,
+	dogma.Command,
 ) error {
 	return nil
 }
 
 // TimeoutHint returns a duration that is suitable for computing a deadline
 // for the handling of the given message by this handler.
-func (FirstIntegrationHandler) TimeoutHint(m dogma.Message) time.Duration {
+func (FirstIntegrationHandler) TimeoutHint(dogma.Message) time.Duration {
 	return 0
 }